pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rosac"
version = "1.0.0"
description = "RouterOS Artifact Collector -- forensic triage for Mikrotik RouterOS devices"
requires-python = ">=3.11"
dependencies = [
    "fabric>=3.2",
    "paramiko>=3.4",
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = ["librouteros>=3.2"]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
]

[project.scripts]
rosac = "rosac.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rosac"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=rosac --cov-report=term-missing"
markers = ["integration: requires live RouterOS device or mock SSH server"]
